server: merge auth services into one fairing
